+++ CHANGELOG.md
@@ -7,6 +7,7 @@ UPGRADE NOTES:
 BUG FIXES:
 
 - Variables with validation no longer interfere with the destroy process ([#3131](https://github.com/opentofu/opentofu/pull/3131))
+- Fixed crash when processing multiple deprecated marks on a complex object ([#3105](https://github.com/opentofu/opentofu/pull/3105))
 
 ## 1.10.5